auto play in all minigames without specific command
